requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
Bcrypt-Flask==1.0.2
python-dotenv==0.19.0
pymysql==1.1.1
Werkzeug==3.1.3
click~=8.2.0
SQLAlchemy~=2.0.41
WTForms~=3.2.1
Flask-WTF~=1.2.2